+++ app/controllers/api/v1/overland/batches_controller.rb
@@ -2,9 +2,10 @@
 
 class Api::V1::Overland::BatchesController < ApplicationController
   skip_forgery_protection
+  before_action :authenticate_api_key
 
   def create
-    Overland::BatchCreatingJob.perform_later(batch_params)
+    Overland::BatchCreatingJob.perform_later(batch_params, current_api_user.id)
 
     render json: { result: 'ok' }, status: :created
   end
@@ -12,6 +13,6 @@ class Api::V1::Overland::BatchesController < ApplicationController
   private
 
   def batch_params
-    params.permit(locations: [:type, geometry: {}, properties: {}], batch: {})
+    params.permit(locations: [:type, { geometry: {}, properties: {} }], batch: {})
   end
 end