+++ app/controllers/api/v1/points_controller.rb
@@ -1,19 +1,15 @@
 # frozen_string_literal: true
 
+# TODO: Deprecate in 1.0
+
 class Api::V1::PointsController < ApplicationController
   skip_forgery_protection
 
   def create
+    Rails.logger.info 'This endpoint will be deprecated in 1.0. Use /api/v1/owntracks/points instead'
     Owntracks::PointCreatingJob.perform_later(point_params)
 
     render json: {}, status: :ok
-  end
-
-  def destroy
-    @point = Point.find(params[:id])
-    @point.destroy
-
-    head :no_content
   end
 
   private