+++ app/jobs/owntracks/point_creating_job.rb
@@ -3,9 +3,10 @@
 class Owntracks::PointCreatingJob < ApplicationJob
   queue_as :default
 
-  def perform(point_params)
+  # TODO: after deprecation of old endpoint, make user_id required
+  def perform(point_params, user_id = nil)
     parsed_params = OwnTracks::Params.new(point_params).call
 
-    Point.create(parsed_params)
+    Point.create!(parsed_params.merge(user_id:))
   end
 end