+++ spec/models/point_spec.rb
@@ -1,8 +1,11 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
 RSpec.describe Point, type: :model do
   describe 'associations' do
     it { is_expected.to belong_to(:import).optional }
+    it { is_expected.to belong_to(:user).optional }
   end
 
   describe 'validations' do