+++ spec/models/user_spec.rb
@@ -7,6 +7,7 @@ RSpec.describe User, type: :model do
     it { is_expected.to have_many(:imports).dependent(:destroy) }
     it { is_expected.to have_many(:points).through(:imports) }
     it { is_expected.to have_many(:stats) }
+    it { is_expected.to have_many(:tracked_points).class_name('Point').dependent(:destroy) }
   end
 
   describe 'callbacks' do