+++ spec/requests/api/v1/overland/batches_spec.rb
@@ -9,16 +9,28 @@ RSpec.describe 'Api::V1::Overland::Batches', type: :request do
     let(:json) { JSON.parse(file.read) }
     let(:params) { json }
 
-    it 'returns http success' do
-      post '/api/v1/overland/batches', params: params
+    context 'with invalid api key' do
+      it 'returns http unauthorized' do
+        post '/api/v1/overland/batches', params: params
 
-      expect(response).to have_http_status(:created)
+        expect(response).to have_http_status(:unauthorized)
+      end
     end
 
-    it 'enqueues a job' do
-      expect do
-        post '/api/v1/overland/batches', params: params
-      end.to have_enqueued_job(Overland::BatchCreatingJob)
+    context 'with valid api key' do
+      let(:user) { create(:user) }
+
+      it 'returns http success' do
+        post "/api/v1/overland/batches?api_key=#{user.api_key}", params: params
+
+        expect(response).to have_http_status(:created)
+      end
+
+      it 'enqueues a job' do
+        expect do
+          post "/api/v1/overland/batches?api_key=#{user.api_key}", params: params
+        end.to have_enqueued_job(Overland::BatchCreatingJob)
+      end
     end
   end
 end