+++ spec/requests/settings_spec.rb
@@ -40,4 +40,32 @@ RSpec.describe 'Settings', type: :request do
       end
     end
   end
+
+  describe 'POST /generate_api_key' do
+    context 'when user is not signed in' do
+      it 'redirects to the sign in page' do
+        post '/settings/generate_api_key'
+
+        expect(response).to redirect_to(new_user_session_path)
+      end
+    end
+
+    context 'when user is signed in' do
+      let(:user) { create(:user) }
+
+      before do
+        sign_in user
+      end
+
+      it 'generates an API key for the user' do
+        expect { post '/settings/generate_api_key' }.to change { user.reload.api_key }
+      end
+
+      it 'redirects back' do
+        post '/settings/generate_api_key'
+
+        expect(response).to redirect_to(root_path)
+      end
+    end
+  end
 end